backend/requirements.txt
# ── Web Framework ────────────────────────────
fastapi
uvicorn

# ── Database ─────────────────────────────────
motor

# ── Data Validation & Serialisation ──────────
pydantic
pydantic-core
email-validator

# ── Exchange Data Ingestion ──────────────────
ccxt

# ── Real-time ────────────────────────────────
websockets

# ── Quant & Analytics ─────────────────────────
pandas

# ── Configuration ────────────────────────────
python-dotenv
cryptography
